Restructure `command-parser`
