+++ command-parser/src/builder/mod.rs
@@ -1,0 +1,3 @@
+mod command;
+
+pub use command::CommandBuilder;
